+++ auth-service/src/main/java/com/sobok/authservice/auth/controller/AuthFeignController.java
@@ -3,7 +3,7 @@ package com.sobok.authservice.auth.controller;
 import com.sobok.authservice.auth.dto.response.AuthLoginResDto;
 import com.sobok.authservice.auth.dto.response.AuthRiderInfoResDto;
 import com.sobok.authservice.auth.dto.response.OauthResDto;
-import com.sobok.authservice.auth.service.auth.AuthService;
+import com.sobok.authservice.auth.service.auth.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
@@ -18,13 +18,15 @@ import java.util.List;
 public class AuthFeignController {
 
     private final AuthService authService;
+    private final InfoService infoService;
+    private final StatusService statusService;
 
     /**
      * 라이더 활성화
      */
     @PutMapping("/active-rider")
     public ResponseEntity<Void> activeRider(@RequestParam Long authId) {
-        authService.activeRider(authId);
+        statusService.activeRider(authId);
         return ResponseEntity.ok().build();
     }
 
@@ -33,20 +35,19 @@ public class AuthFeignController {
      */
     @GetMapping("/auth/info")
     public ResponseEntity<AuthRiderInfoResDto> getAuthInfo(@RequestParam Long authId) {
-        return ResponseEntity.ok(authService.getRiderAuthInfo(authId));
+        return ResponseEntity.ok(infoService.getRiderAuthInfo(authId));
     }
 
     //oauthId가 존재할 때 - oauthId로 authId 찾기
     @GetMapping("/findByOauthId")
     OauthResDto authIdById(@RequestParam("id") Long id) {
-        OauthResDto byOauthId = authService.findByOauthId(id);
+        OauthResDto byOauthId = infoService.findByOauthId(id);
         log.info("authId: {}", byOauthId);
         return byOauthId;
     }
 
     @GetMapping("/social-token")
     AuthLoginResDto socialToken(@RequestParam("authId") Long id) {
-        log.info("여기는 토큰생성하는길");
         return authService.socialLoginToken(id);
     }
 
@@ -55,12 +56,12 @@ public class AuthFeignController {
      */
     @GetMapping("/auth/login-id")
     public ResponseEntity<String> getLoginId(@RequestParam Long authId) {
-        String loginId = authService.getLoginIdByAuthId(authId);
+        String loginId = infoService.getLoginIdByAuthId(authId);
         return ResponseEntity.ok(loginId);
     }
 
     @GetMapping("/get-rider-inactive")
     List<Long> getInactiveRidersInfo() {
-        return authService.getInactiveRidersInfo();
+        return infoService.getInactiveRidersInfo();
     }
 }
